binary_serde

Struct BinaryDeserializerFromBuf

pub struct BinaryDeserializerFromBuf<'a> { /* private fields */ }

a deserializer which deserializes values from a buffer.

Implementations

impl<'a> BinaryDeserializerFromBuf<'a>

pub fn new(buf: &'a [u8], endianness: Endianness) -> Self

creates a new deserializer which deserializes values from the given buffer using the given endianness.

pub fn deserialize<T: BinarySerde>(&mut self) -> Result<T, DeserializeError>

deserializes a value of type T from the current position in the buffer, and advances the position accordingly.

Panics

this function panics if the deserialization exceeds the bounds of the buffer.

pub fn position(&self) -> usize

returns the current position of this deserializer in the buffer.

pub fn set_position(&mut self, new_position: usize)

sets the position of this deserializer in the buffer.

pub fn move_forwards(&mut self, amount: usize)

moves this deserializer’s position forwards according to the given amount.

pub fn move_backwards(&mut self, amount: usize)

moves this deserializer’s position backwards according to the given amount.

pub fn set_endianness(&mut self, new_endianness: Endianness)

sets the endianness of this deserializer

pub fn endianness(&self) -> Endianness

returns the endianness of this deserializer.

pub fn buf(&self) -> &'a [u8]

returns a reference to the underlying buffer of this deserializer